+++ tsconfig.json
@@ -22,7 +22,7 @@
 			"@/*": ["./src/*"],
 			"@components/*": ["./src/app/_components/*"],
 			"@features/*": ["./src/app/_features/*"],
-			"@libs/*": ["./src/libs/*"]
+			"@libs/*": ["./src/app/_libs/*"]
 		}
 	},
 	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
